removing BaseKNN since I have no convinced that it is useful to have. I do agree it is weird to be based on a classifier and then be used for regressor! So likely this is a better design
