+++ river/neighbors/__init__.py
@@ -4,12 +4,12 @@ Also known as *lazy* methods. In these methods, generalisation of the training d
 until a query is received.
 
 """
-from .base_neighbors import BaseKNN
 from .knn_classifier import KNNClassifier
 from .knn_regressor import KNNRegressor
+from .neighbors import NearestNeighbors
 
 __all__ = [
-    "BaseKNN",
+    "NearestNeighbors",
     "KNNClassifier",
     "KNNRegressor",
 ]
