+++ river/neighbors/knn_classifier.py
@@ -1,12 +1,11 @@
 from river import base, utils
 
-from .base_neighbors import BaseKNN
-from .neighbors import DistanceFunc
+from .neighbors import DistanceFunc, MinkowskiNeighbors
 
 __all__ = ["KNNClassifier"]
 
 
-class KNNClassifier(BaseKNN, base.Classifier):
+class KNNClassifier(base.Classifier):
     """
     K-Nearest Neighbors (KNN) for classification.
 
@@ -83,17 +82,26 @@ class KNNClassifier(BaseKNN, base.Classifier):
         distance_func: DistanceFunc = None,
         softmax: bool = False,
     ):
-        super().__init__(
-            n_neighbors=n_neighbors,
-            window_size=window_size,
-            min_distance_keep=min_distance_keep,
-            distance_func=distance_func,
-        )
         self.weighted = weighted
         self.cleanup_every = cleanup_every
         self.classes = set()
         self.softmax = softmax
         self._cleanup_counter = cleanup_every
+
+        self.n_neighbors = n_neighbors
+        self.window_size = window_size
+        self.min_distance_keep = min_distance_keep
+        self.distance_func = distance_func
+        self.nn = MinkowskiNeighbors(
+            window_size=window_size,
+            distance_func=distance_func,
+            min_distance_keep=min_distance_keep,
+            n_neighbors=n_neighbors,
+        )
+
+    @property
+    def _multiclass(self):
+        return True
 
     def clean_up_classes(self) -> "KNNClassifier":
         """
@@ -132,6 +140,15 @@ class KNNClassifier(BaseKNN, base.Classifier):
                 self._cleanup_counter = self.cleanup_every
 
         return self
+
+    def predict_one(self, x: dict):
+        """Predict the label of a set of features `x`.
+        Parameters:
+            x: A dictionary of features.
+        Returns:
+            The neighbors
+        """
+        return self.predict_proba_one(x)
 
     def predict_proba_one(self, x):
         """Predict the class of a set of features `x`.
